requirements.txt
flask
numpy
imageio
opencv-python
scipy
confuse
pillow
pywin32